add remaining NonSiloTests to NonSiloTests in both vNext and main solution
