+++ test/NonSiloTests/DeploymentTests/Configuration/TestAbsolutePaths/OrleansConfiguration.xml
