+++ test/NonSiloTests/General/InterfaceRules.cs
