+++ test/NonSiloTests/General/InternalSerializationTests.cs
