+++ test/NonSiloTests/General/RingTests_Standalone.cs
