+++ test/NonSiloTests/Serialization/SerializerGenerationTests.cs
