+++ test/TesterInternal/TesterInternal.csproj
@@ -60,7 +60,6 @@
     <Compile Include="HostedTestClusterBase.cs" />
     <Compile Include="MembershipTests\DynamoDBMembershipTableTest.cs" />
     <Compile Include="MembershipTests\PostgreSqlMembershipTableTests.cs" />
-    <Compile Include="OrleansRuntime\Streams\SubscriptionMarkerTests.cs" />
     <Compile Include="RemindersTest\DynamoDBRemindersTableTests.cs" />
     <Compile Include="RemindersTest\PostgreSqlRemindersTableTests.cs" />
     <Compile Include="SqlStatisticsPublisherTests\PostgreSqlStatisticsPublisherTests.cs" />
@@ -101,32 +100,20 @@
     <Compile Include="StorageTests\Relational\TestDataSets\TestStateGeneric1.cs" />
     <Compile Include="StorageTests\Relational\TestEnvironmentInvariant.cs" />
     <Compile Include="StreamingTests\AQStreamingTests.cs" />
-    <Compile Include="AssemblyLoaderTests.cs" />
-    <Compile Include="AsynchAgentRestartTest.cs" />
-    <Compile Include="AsyncPipelineTests.cs" />
-    <Compile Include="Async_AsyncExecutorWithRetriesTests.cs" />
-    <Compile Include="Async_TimingTests.cs" />
     <Compile Include="ClientAddressableTests.cs" />
     <Compile Include="ClientInitTests.cs" />
     <Compile Include="CodeGeneratorTests_RequiringSilo.cs" />
-    <Compile Include="CodeGeneratorTests.cs" />
     <Compile Include="ConcurrencyTests.cs" />
     <Compile Include="CounterControlProgTests.cs" />
-    <Compile Include="CounterTests.cs" />
     <Compile Include="DeadlockDetectionTests.cs" />
     <Compile Include="EchoTaskGrainTests.cs" />
     <Compile Include="ErrorGrainTest.cs" />
-    <Compile Include="ErrorInjectionStorageProvider.cs" />
-    <Compile Include="GatewaySelectionTest.cs" />
-    <Compile Include="General\RingTests_Standalone.cs" />
     <Compile Include="General\ConsistentRingProviderTests_Silo.cs" />
     <Compile Include="General\BootstrapProviderTests.cs" />
-    <Compile Include="General\CounterStatisticTests.cs" />
     <Compile Include="General\ElasticPlacementTest.cs" />
     <Compile Include="General\GrainPlacementTests.cs" />
     <Compile Include="General\LoadSheddingTest.cs" />
     <Compile Include="General\LoggerTest.cs" />
-    <Compile Include="General\LruTest.cs" />
     <Compile Include="General\RequestContextTest.cs" />
     <Compile Include="General\TestingBootstrapProviders.cs" />
     <Compile Include="GeoClusterTests\AzureGossipTableTests.cs" />
@@ -143,14 +130,11 @@
     <Compile Include="OrleansRuntime\Streams\PooledQueueCacheTests.cs" />
     <Compile Include="SqlStatisticsPublisherTests\SqlServerStatisticsPublisherTests.cs" />
     <Compile Include="SqlStatisticsPublisherTests\MySqlStatisticsPublisherTests.cs" />
-    <Compile Include="General\Identifiertests.cs" />
-    <Compile Include="General\InterfaceRules.cs" />
     <Compile Include="GrainReferenceCacheTests.cs" />
     <Compile Include="GrainReferenceCastTests.cs" />
     <Compile Include="GrainReferenceTest.cs" />
     <Compile Include="MembershipTests\MySqlMembershipTableTests.cs" />
     <Compile Include="LocalErrorGrain.cs" />
-    <Compile Include="ManagementAgentTests.cs" />
     <Compile Include="MockStatsCollectors.cs" />
     <Compile Include="ReentrancyTests.cs" />
     <Compile Include="SchedulerTests\STSchedulerLongTurnTest.cs" />
@@ -164,9 +148,6 @@
     <Compile Include="StorageTests\LocalStoreTests.cs" />
     <Compile Include="MultifacetGrainTest.cs" />
     <Compile Include="OrleansHostProgTests.cs" />
-    <Compile Include="OrleansRuntime\AsyncSerialExecutorTests.cs" />
-    <Compile Include="OrleansRuntime\ExceptionsTests.cs" />
-    <Compile Include="OrleansRuntime\Streams\BestFitBalancerTests.cs" />
     <Compile Include="PolymorphicInterfaceTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="..\..\src\Build\GlobalAssemblyInfo.cs">
@@ -176,8 +157,6 @@
     <Compile Include="RemindersTest\MySqlRemindersTableTests.cs" />
     <Compile Include="RemindersTest\ReminderTableTestsBase.cs" />
     <Compile Include="RemindersTest\SqlServerRemindersTableTests.cs" />
-    <Compile Include="General\InternalSerializationTests.cs" />
-    <Compile Include="Serialization\SerializerGenerationTests.cs" />
     <Compile Include="StatsTests.cs" />
     <Compile Include="StorageTests\AzureTableDataManagerStressTests.cs" />
     <Compile Include="StorageTests\AzureTableErrorCodeTests.cs" />
@@ -210,7 +189,6 @@
     <Compile Include="StreamingTests\StreamTestUtils.cs" />
     <Compile Include="StreamProvidersTests.cs" />
     <Compile Include="TestingHost\TestingSiloHostTests.cs" />
-    <Compile Include="TestUtils.cs" />
     <Compile Include="TimeoutTests.cs" />
     <Compile Include="TimerOrleansTest.cs" />
     <Compile Include="TimerTests\ReminderTests_Azure_Standalone.cs" />
@@ -392,7 +370,6 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
   </ItemGroup>
-  <ItemGroup />
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <PropertyGroup Condition=" '$(OS)' == 'Windows_NT' ">
     <!-- PostBuild.cmd script is platform specific, so for now we only want to run it on Windows -->
