+++ vNext/test/TestGrainInterfaces/Properties/AssemblyInfo.cs
@@ -5,8 +5,8 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("TestExtensions")]
-[assembly: AssemblyDescription("")]
+[assembly: AssemblyTitle("UnitTests.GrainInterfaces")]
+[assembly: AssemblyDescription("UnitTests.GrainInterfaces")]
 [assembly: AssemblyConfiguration("")]
 
 // Setting ComVisible to false makes the types in this assembly not visible 
@@ -15,6 +15,6 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("8fd242b4-eda9-42cd-ba39-e410b98add26")]
+[assembly: Guid("a8df0713-6001-4830-af7e-dcb114bfb340")]
 
-[assembly: InternalsVisibleTo("TesterInternal")]
+[assembly: InternalsVisibleTo("Orleans.NonSiloTests")]
